Fix issues with potentially misaligned storage in bloom_filter_ref (#660)
The current design of `cuco::bloom_filter_ref` might lead to a compiler
segfault if the storage pointer passed to the ctor is misaligned.

This PR fixes this issue by introducing an opaque strong type
`filter_block_type` that comes with the correct alignment, which makes
it easier for the user to allocate properly aligned storage and makes
the alignment check that fails through a `__trap()` mechanism (which
likely caused the segfault in the first place) obsolete.